Generated initial compilers.yaml file for E X P A N S E
This compilers.yaml file was generated by running the 'spack compilers'
command and only contains the core system gcc compiler set [gcc (GCC)
8.3.1 20190507 (Red Hat 8.3.1-4)] that was available on Expanse at the
time of this commit. The only custom modifications made to the file were
the addition of the compiler optimization flags as shown here.

A copy of this file will also serve as the initial compilers.yaml file
for the sdsc/expanse/0.15.4/gpu Spack instance, whose configuration will
also be documented here on the sdsc-0.15.4 Spack deployment branch in
future commits.