Created initial packages.yaml file for E X P A N S E
To date, the packages.yaml file has been primarily used on Expanse to
specify which external system packages Spack will build against. This
first instantiation only specifies the default system openssl for core
Spack compiler [gcc (GCC) 8.3.1 20190507 (Red Hat 8.3.1-4)]. Additional
external packages will be added as this deployment branch is created.

We have also used packages.yaml on Expanse to restrict licensed software
to specific Linux groups as well. These permissions will be also be
added to this packages.yaml as the deployment branch is created. e.g.,
when such restricted packages are added to be tracked in the repo.