Deploy gcc@10.2.0 compilers with AMD Rome (zen2) support
gcc@10.2.0 was the first version of the gcc compilers to support
optimizations for the AMD Rome (zen2) architecture. Note, however,
gcc@10.2.0 was not available and/or supported by Spack at the v0.15.4
release tag, which is why this initial Spack deployment branch begins
at an intermediate commit (8edb31e93). We do not to expect to create
future Spack deployment branches from similar intermediate commits. This
was a special case.