Move spec variable declarations after sourcing target instance
Moved all environment variables associated with specifying a Spack
package's install spec for a target instance in each Slurm-based build +
install job script to be declared after the target instance has been
sourced into the job's environment. This is important as we will often
use spack commands to construct explicit SPACK_DEPENDENCIES in order to
force the re-use of existing software package variants already installed
within the target instance. The reason for this is that the Spack
concretizer does not yet do a good job of attempting to re-use existing
software packages for a new package's dependencies when the existing
software was not built and installed with the default Spack spec. i.e.,
any Spack package installed with non-default variant options will likely
not be used to satisfy a new package's dependencies when its spec is
concretized, unless explicitly set in SPACK_DEPENDENCIES.