Add init gcc@10.2.0 package dependency chain for expanse/0.15.4/cpu
This set of Slurm-based build + install job scripts comprise an initial
Spack package dependency chain used to populate the expanse/0.15.4/cpu
instance. It begins by installing lmod within the core gcc@8.3.1
compiler environment to support the use of lmod software modules with
Spack and then bootstraps the installation of the instance's first
internal compiler, gcc@10.2.0, which is then used to begin building out
the entire software stack associated with this compiler. At this time,
this initial package dependency chain focuses on deploying standard
build tools and utilities like cmake, programming support for languages
like perl, python, java, and go, and simple data visualiztion and image
manipulation tools like gnuplot and imagemagick. The next set of
packages to be added to the chain will be serial and/or shared memory
versions of core math and scientific software libraries that will serve
as underlying dependencies for many of the scientific applications to be
added later on.