Set harfbuzz variant default to be +graphite2 in sdsc package repo
texlive requires the harfbuzz variant to include +graphite2, but the
builtin harfbuzz package sets ~graphite2 as its default. This change is
important to satisfy the dependencies for a number of visualization
packages that may include support for latex formatting of text within
the figures and plots they generate. e.g., py-matplotlib +latex.

Also renamed the OpenBLAS + OpenMP build + install script and appended
the 'openmp' suffix to the FFTW and OpenBLAS module naming schemes within
the modules.yaml file.