Modify cairo package to set all variant default=True
This modification was made to the cairo package and placed in the sdsc
package repository to help satisfy and resolve dependency conflicts with
R plotting packages that will appear further down the gcc@10.2.0
dependency chain. It should also help minimize the number of cairo
versions installed within the gcc@10.2.0 chain.