Remove cairo from sdsc package repo
Setting all of cairo's default variant values to True was found to cause
ghostscript to fail concretization with 'Error: maximum recursion depth
exceeded'. No other remedy was found to resolve the issue and, as such,
the only temporary solution is to remove the modified package from the
sdsc package repository. We will have to address downstream R dependency
issue later.