Turn on dependency autoloading for ^python in modules.yaml
Turned on dependency autoloading in the modules.yaml file for all
packages that depend on python in the expanse/0.15.4/cpu instance.
During the next iteration of the bootstrapping process here, this
autoloading feature will be tested to determine if it should be applied
to all module-loaded Spack packages. One of the fundamental questions is
whether or not the Spack package dependencies without a module will also
be autoloaded in the background as well.