Add lustre, rdma-core, and slurm as external packages
Added lustre, rdma-core, and slurm as external packages in the
packages.yaml file for expanse/0.15.4/cpu. These external packages are
required to integrate Lustre, Infiniband, and Slurm support with the
different MPI distributions on Expanse. The first MPI distribution to be
installed for the gcc@10.2.0 dependency chain will be OpenMPI version 4.0.5.